backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
yt-dlp==2023.12.30
openai-whisper==20231117
anthropic==0.39.0
torch==2.1.0
stripe==7.7.0
supabase==2.3.0
python-dotenv==1.0.0
pyjwt==2.8.0
gunicorn==21.2.0
reportlab==4.0.7
python-docx==1.1.0
requests==2.31.0

